% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Criteria.R
\name{Criteria}
\alias{Criteria}
\title{AUC and BS}
\usage{
Criteria(s, t, Survt, CR, P, cause)
}
\arguments{
\item{s}{the landmark time for prediction}

\item{t}{the window of prediction for prediction}

\item{Survt}{the survival time}

\item{CR}{the indicator for competing risks}

\item{P}{the risk prediction}

\item{cause}{the main cause for prediction}
}
\value{
\itemize{
\item AUC The values of AUC from timeROC package
\item BS The values of BS from available documents
\item Cri A summary of AUC and BS
}
}
\description{
Compute the iid-representation of areas under time-dependent ROC curves (AUC) and Brier score (BS)
}
\details{
The Brier score and AUC calculations utilize the available functions within the timeROC package, with added flexibility in the function arguments. Alongside the obtained results, a summary including estimates and standard deviations is provided for enhanced interpretation.
}
\examples{
rm(list=ls())
library(Landmarking)
library(pec)
data(data)
set.seed(2)
INDTRAIN <- sample(unique(data$id), 0.5 * length(unique(data$id)))
INDVALID <- unique(data$id)[-INDTRAIN]
########################
data_model_landmark_LOCF <-return_ids_with_LOCF(
  data_long = data,
  individual_id = "id",
  covariates =c("Y1","Y2"),
  covariates_time =c(rep("obstime", 2)),
  x_L =0.2
)

data_t=data_model_landmark_LOCF[(data_model_landmark_LOCF$id \%in\% c(INDTRAIN)),]

data_model_landmark_LOCF1 <-fit_LOCF_landmark(
  data_long = data_t,
  x_L = 0.2,
  x_hor = 0.5,
    covariates =c("Y1","Y2"),
    covariates_time =c(rep("obstime", 2)),
  individual_id = "id",
  event_time = "Time",
  event_status = "Censoring",
  survival_submodel = "standard_cox"
)
data_v=data_model_landmark_LOCF[!(data_model_landmark_LOCF$id \%in\% c(INDVALID)),]

AA=predict(object=data_model_landmark_LOCF1,x_L=0.2,
           x_hor =0.5,newdata=data_v)

DP=AA[,4]

AAA=data.frame(AA$id,data_v[!duplicated(data_v[,2]),],DP)


Criteria(s=0.1, t=0.5, Survt=AAA$Time, CR=AAA$Censoring, P=DP, cause=1)
}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}
}
